Move assert_ranges! into macros::clamp
