+++ palette/src/lib.rs
@@ -325,162 +325,6 @@ pub use matrix::Mat3;
 #[allow(deprecated)]
 pub use relative_contrast::{contrast_ratio, RelativeContrast};
 
-//Helper macro for checking ranges and clamping.
-#[cfg(test)]
-macro_rules! assert_ranges {
-    (@make_tuple $first:pat, $next:ident,) => (($first, $next));
-
-    (@make_tuple $first:pat, $next:ident, $($rest:ident,)*) => (
-        assert_ranges!(@make_tuple ($first, $next), $($rest,)*)
-    );
-
-    (
-        $ty:ident < $($ty_params:ty),+ >;
-        clamped {$($clamped:ident: $clamped_from:expr => $clamped_to:expr),+}
-        clamped_min {$($clamped_min:ident: $clamped_min_from:expr => $clamped_min_to:expr),*}
-        unclamped {$($unclamped:ident: $unclamped_from:expr => $unclamped_to:expr),*}
-    ) => (
-        {
-            use core::iter::repeat;
-            use crate::{Clamp, IsWithinBounds};
-
-            {
-                print!("checking below clamp bounds... ");
-                $(
-                    let from = $clamped_from;
-                    let to = $clamped_to;
-                    let diff = to - from;
-                    let $clamped = (1..11).map(|i| from - (i as f64 / 10.0) * diff);
-                )+
-
-                $(
-                    let from = $clamped_min_from;
-                    let to = $clamped_min_to;
-                    let diff = to - from;
-                    let $clamped_min = (1..11).map(|i| from - (i as f64 / 10.0) * diff);
-                )*
-
-                $(
-                    let from = $unclamped_from;
-                    let to = $unclamped_to;
-                    let diff = to - from;
-                    let $unclamped = (1..11).map(|i| from - (i as f64 / 10.0) * diff);
-                )*
-
-                for assert_ranges!(@make_tuple (), $($clamped,)+ $($clamped_min,)* $($unclamped,)* ) in repeat(()) $(.zip($clamped))+ $(.zip($clamped_min))* $(.zip($unclamped))* {
-                    let color: $ty<$($ty_params),+> = $ty {
-                        $($clamped: $clamped.into(),)+
-                        $($clamped_min: $clamped_min.into(),)*
-                        $($unclamped: $unclamped.into(),)*
-                        ..$ty::default() //This prevents exhaustiveness checking
-                    };
-
-                    let clamped = color.clamp();
-
-                    let expected: $ty<$($ty_params),+> = $ty {
-                        $($clamped: $clamped_from.into(),)+
-                        $($clamped_min: $clamped_min_from.into(),)*
-                        $($unclamped: $unclamped.into(),)*
-                        ..$ty::default() //This prevents exhaustiveness checking
-                    };
-
-                    assert!(!color.is_within_bounds());
-                    assert_relative_eq!(clamped, expected);
-                }
-
-                println!("ok")
-            }
-
-            {
-                print!("checking within clamp bounds... ");
-                $(
-                    let from = $clamped_from;
-                    let to = $clamped_to;
-                    let diff = to - from;
-                    let $clamped = (0..11).map(|i| from + (i as f64 / 10.0) * diff);
-                )+
-
-                $(
-                    let from = $clamped_min_from;
-                    let to = $clamped_min_to;
-                    let diff = to - from;
-                    let $clamped_min = (0..11).map(|i| from + (i as f64 / 10.0) * diff);
-                )*
-
-                $(
-                    let from = $unclamped_from;
-                    let to = $unclamped_to;
-                    let diff = to - from;
-                    let $unclamped = (0..11).map(|i| from + (i as f64 / 10.0) * diff);
-                )*
-
-                for assert_ranges!(@make_tuple (), $($clamped,)+ $($clamped_min,)* $($unclamped,)* ) in repeat(()) $(.zip($clamped))+ $(.zip($clamped_min))* $(.zip($unclamped))* {
-                    let color: $ty<$($ty_params),+> = $ty {
-                        $($clamped: $clamped.into(),)+
-                        $($clamped_min: $clamped_min.into(),)*
-                        $($unclamped: $unclamped.into(),)*
-                        ..$ty::default() //This prevents exhaustiveness checking
-                    };
-
-                    let clamped = color.clamp();
-
-                    assert!(color.is_within_bounds());
-                    assert_relative_eq!(clamped, color);
-                }
-
-                println!("ok")
-            }
-
-            {
-                print!("checking above clamp bounds... ");
-                $(
-                    let from = $clamped_from;
-                    let to = $clamped_to;
-                    let diff = to - from;
-                    let $clamped = (1..11).map(|i| to + (i as f64 / 10.0) * diff);
-                )+
-
-                $(
-                    let from = $clamped_min_from;
-                    let to = $clamped_min_to;
-                    let diff = to - from;
-                    let $clamped_min = (1..11).map(|i| to + (i as f64 / 10.0) * diff);
-                )*
-
-                $(
-                    let from = $unclamped_from;
-                    let to = $unclamped_to;
-                    let diff = to - from;
-                    let $unclamped = (1..11).map(|i| to + (i as f64 / 10.0) * diff);
-                )*
-
-                for assert_ranges!(@make_tuple (), $($clamped,)+ $($clamped_min,)* $($unclamped,)* ) in repeat(()) $(.zip($clamped))+ $(.zip($clamped_min))* $(.zip($unclamped))* {
-                    let color: $ty<$($ty_params),+> = $ty {
-                        $($clamped: $clamped.into(),)+
-                        $($clamped_min: $clamped_min.into(),)*
-                        $($unclamped: $unclamped.into(),)*
-                        ..$ty::default() //This prevents exhaustiveness checking
-                    };
-
-                    let clamped = color.clamp();
-
-                    let expected: $ty<$($ty_params),+> = $ty {
-                        $($clamped: $clamped_to.into(),)+
-                        $($clamped_min: $clamped_min.into(),)*
-                        $($unclamped: $unclamped.into(),)*
-                        ..$ty::default() //This prevents exhaustiveness checking
-                    };
-
-                    assert!(!color.is_within_bounds());
-                    assert_relative_eq!(clamped, expected);
-                }
-
-                println!("ok")
-            }
-        }
-    );
-}
-
 #[macro_use]
 mod macros;
